Something has to be wrong with Brackets
so basically i converted from Notepad++ to Brackets today and the issue
is when i use the "onmouseover" event the image transition is immediate,
but when i use the "onclick" event the transition is smooth. However
when i use the exact same code in Notepad++ either event executes my
function properly.